refactor: refactoring post_board
