+++ src/App.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { BrowserRouter, Route, Routes } from "react-router-dom";
 import { Home } from './ui/home/Home';
 import './App.css';
-import PostBoardScreen from './ui/home/home_main/post_board/PostBoardScreen';
+import PostBoardScreen from './ui/posts/post_board/PostBoardScreen';
 import Posts from './ui/posts/Posts';
 
 function App() {
